requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0

# API and Web
fastapi>=0.70.0
uvicorn>=0.15.0
requests>=2.26.0
aiohttp>=3.8.0
wbdata>=0.7.1

# Data processing
beautifulsoup4>=4.10.0
nltk>=3.6.0
spacy>=3.2.0

# Visualization (for Streamlit dashboard)
streamlit>=1.0.0
plotly>=5.0.0
matplotlib>=3.4.0

# Database and caching
redis>=4.0.0
sqlalchemy>=1.4.0

# Testing
pytest>=6.2.0
pytest-asyncio>=0.16.0

# Utilities
python-dotenv>=0.19.0
pydantic>=1.8.0
PyYAML>=6.0

# Database and caching
redis>=4.0.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.9
pycountry>=22.3.5
